move MdInput files and update imports
