+++ components/ChallengeMaterial/ChallengeMaterial.tsx
@@ -22,7 +22,7 @@ import Modal from 'react-bootstrap/Modal'
 import { SubmissionStatus, useAddCommentMutation } from '../../graphql'
 import DiffView from '../DiffView'
 import { Button } from '../theme/Button'
-import { MdInput } from '../MdInput'
+import MdInput from '../MdInput'
 import { updateCache } from '../../helpers/updateCache'
 import { GlobalContext } from '../../helpers/globalContext'
 import { SubmissionComments } from '../SubmissionComments'
