+++ components/ExerciseReportCard/ExerciseReportCard.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import { DISCORD_PATH } from '../../constants'
 import { FlagExerciseMutation, useFlagExerciseMutation } from '../../graphql'
 import styles from './exerciseReportCard.module.scss'
-import { MdInput } from '../MdInput'
+import MdInput from '../MdInput'
 import NavLink from '../NavLink'
 import QueryInfo from '../QueryInfo'
 import { Button } from '../theme/Button'
