+++ components/ReviewCard.tsx
@@ -19,7 +19,7 @@ import { Button } from './theme/Button'
 import { Text } from './theme/Text'
 import { Accordion } from 'react-bootstrap'
 import Markdown from 'markdown-to-jsx'
-import { MdInput } from './MdInput'
+import MdInput from './MdInput'
 import DiffView from './DiffView'
 import { updateCache } from '../helpers/updateCache'
 import { SelectIteration } from './SelectIteration'
